fixed warnings

+++ cortex-m0plus/1.0.1/cores/cortex-m0plus/efm_lib/LEUARTClass.h
@@ -32,7 +32,7 @@ class LEUARTClass : public HardwareSerial
   LEUARTClass(Leuart *pUart, IRQn_Type dwIrq, uint32_t dwId, RingBuffer *pRx_buffer, RingBuffer *pTx_buffer, uint8_t port, uint8_t txpin, uint8_t rxpin, uint32_t location, uint32_t cmu_clken, uint32_t leuart_clkdiv);
 
   void   begin(const uint32_t dwBaudRate);
-  void   begin(const uint32_t dwBaudRate, uint8_t config);
+  void   begin(const uint32_t dwBaudRate, uint32_t config);
   void   end(void);
   int    available(void);
   int    availableForWrite(void);
@@ -52,7 +52,7 @@ class LEUARTClass : public HardwareSerial
   operator bool() { return true; }; // LEUART0 always active
 
  protected:
-  void init(const uint32_t dwBaudRate, uint8_t config);
+  void init(const uint32_t dwBaudRate, uint32_t config);
 
   RingBuffer *_rx_buffer;
   RingBuffer *_tx_buffer;
